pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=2.18.1",
    "ell-ai>=0.0.13",
    "kaleido>=0.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pyxlsb>=1.0.10",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
constraint-dependencies = ["kaleido!=0.2.1.post1"]
